Prevent forwarder injection after registration (#25)
There were two truthy checks of an item in session storage that was
expected to be a boolean. Instead the item was a string ("true" or
"false"). The checks have been updated to expect a string.

This bug was resulting in the library thinking that registration was
always in progress, once the `REGISTRATION_IN_PROGRESS` variable was
set for the first time.