Ensure all onboarding states are handled
This change makes the default case of `_onMessageFromForwarder` of type `never`,
which ensures that all of the onboarding states must be contained in the `switch`
cases.

When a state is missing:

```bash
$ git d
```
```diff
@@ -82,9 +82,9 @@ export default class Onboarding {

   async _onMessageFromForwarder (event: MessageEvent) {
     switch (this.state) {
-      case ONBOARDING_STATE.RELOADING:
-        console.debug('Ignoring message while reloading');
-        break;
+      // case ONBOARDING_STATE.RELOADING:
+      //   console.debug('Ignoring message while reloading');
+      //   break;
       case ONBOARDING_STATE.NOT_INSTALLED:
         console.debug('Reloading now to register with MetaMask');
         this.state = ONBOARDING_STATE.RELOADING;
```
```bash
$ yarn tsc --project .
```
```
yarn run v1.22.4
$ tsc --project .
src/index.ts:111:42 - error TS2345: Argument of type '"RELOADING"' is not assignable to parameter of type 'never'.

111         this._onMessageUnknownStateError(this.state);
                                             ~~~~~~~~~~

Found 1 error.
```

TypeScript will flag the default state as trying to assign a state to type `never`.