Switch from CircleCI to GitHub Actions
The standard GitHub Actions workflow from `metamask-module-template`
has been used, except that the tests are omitted (because there are no
tests yet), and there is no allow-scripts step yet.

The main difference between the GitHub Actions config and the CircleCI
config is that now we test on Node v14 and v16 in addition to v12, and
we cache dependencies.