Fix import of `bowser` package (#63)
Our import of the `bowser` package was broken because we were importing
all named exports, rather than the default export. The package has no
named exports - only a default import.

This mistake was made because of challenges encountered when trying to
configure the CommonJS and bundled builds. At the time `bowser` didn't
use a `module` field in their `package.json`, so we were importing
their CommonJS bundle. Also, we hadn't at that time hadn't adopted the
practice of using ESModuleInterop for all of our TypeScript packages,
which greatly simplifies this case.

ESModuleInterop is now enabled, and the import has been fixed. The
`@rollup/plugin-commonjs` is no longer needed either, since we're using
the ES6 `bowser` module directly now.

This was mostly fixed already by PR #57. This PR is a follow-up to
standardize the TypeScript config settings (by using `ESModuleInterop`
over `allowSyntheticDefaultImports`), and to ensure the bundle is
correct as well.

Fixes #46